--- configoptions.txt
@@ -1,42 +1,0 @@
-{
-    "__Comment": "Dummy to add comments later on",
-    "Always continuing": [],
-    "Blocked": [],
-    "Board ID": "",
-    "Version": "1.6",
-    "Doing": [],
-    "Done": [],
-    "Epics": [],
-    "List with Epics Done": [],
-    "Email address to send": "",
-    "Excel Outputfile": "",
-    "Google Spreadsheet ID": "",
-    "Google sheetname for all Trello data": "",
-    "Google sheetname for timeline": "",
-    "JSON file from Google": "",
-    "Maximum days a card can be in Done": "",
-    "MySQL table for all Trello data": "",
-    "MySQL table for timeline": "",
-    "MySQL Database name": "",
-    "Lists to ignore for Dashboard": [],
-    "Not Started": [],
-    "Script options": {
-        "Clean the Done lists": false,
-        "Output a timeline to MySQL server": false,
-        "Output all data to MySQL server": false,
-        "Output a timeline to Excel": false,
-        "Output a timeline to Google Sheet": false,
-        "Output all data to Excel": false,
-        "Output all data to Google Sheets": false,
-        "Remove members from Done and Archived cards": false,
-        "Calculate hours": false
-    },
-    "Subject of Email": "",
-    "List for hours": "",
-    "Custom Field with hours": "",
-    "Custom Field for Starting date": "",
-    "Custom Field for Ending date": "",
-    "Custom Field for Categories": "",
-    "Custom Field for Person": "" 
-
-}
