--- credentialsoptions.txt
@@ -1,14 +1,0 @@
-{
-    "__Comment": "Dummy to comment on file later on",
-    "API key": "",
-    "API token": "",
-    "Gmail Address": "",
-    "Gmail password": "",
-    "MySQL password": "",
-    "MySQL port": "",
-    "MySQL server": "",
-    "MySQL user": "",
-    "Version": "1.3",
-    "Username for Dash": "",
-    "Password for Dash": ""
-}
